backend/requirements.txt
fastapi
google-genai
python-dotenv
python-multipart
uvicorn
